% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_american.R
\name{scale_color_american}
\alias{scale_color_american}
\title{Scale Colors as American University}
\usage{
scale_color_american(palette = c("american"), alpha = 1, ...)
}
\description{
Scale Colors as American University
}
