% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_berkeley.R
\name{scale_color_berkeley}
\alias{scale_color_berkeley}
\title{Scale Colors as UC Berkeley}
\usage{
scale_color_berkeley(palette = c("berkeley"), alpha = 1, ...)
}
\description{
Scale Colors as UC Berkeley
}
