% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_brownu.R
\name{scale_color_brownu}
\alias{scale_color_brownu}
\title{Scale Colors as Brown University}
\usage{
scale_color_brownu(palette = c("brownu"), alpha = 1, ...)
}
\description{
Scale Colors as Brown University
}
