% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_georgetown.R
\name{scale_color_georgetown}
\alias{scale_color_georgetown}
\title{Scale Colors as Georgetown University}
\usage{
scale_color_georgetown(palette = c("georgetown"), alpha = 1, ...)
}
\description{
Scale Colors as Georgetown University
}
