% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_harvard.R
\name{scale_color_harvard}
\alias{scale_color_harvard}
\title{Scale Colors as Harvard}
\usage{
scale_color_harvard(palette = c("harvard"), alpha = 1, ...)
}
\description{
Scale Colors as Harvard
}
