% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_nyu.R
\name{scale_color_nyu}
\alias{scale_color_nyu}
\title{Scale Colors as New York University}
\usage{
scale_color_nyu(palette = c("nyupal"), alpha = 1, ...)
}
\description{
Scale Colors as New York University
}
