% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_stanford.R
\name{scale_color_stanford}
\alias{scale_color_stanford}
\title{Scale Colors as Stanford University}
\usage{
scale_color_stanford(palette = c("stanford"), alpha = 1, ...)
}
\description{
Scale Colors as Stanford University
}
