% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_uva.R
\name{scale_color_uva}
\alias{scale_color_uva}
\title{Scale Colors as University of Virginia}
\usage{
scale_color_uva(palette = c("uvapal"), alpha = 1, ...)
}
\description{
Scale Colors as University of Virginia
}
