% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_vatech.R
\name{scale_color_vatech}
\alias{scale_color_vatech}
\title{Scale Colors as Virginia Polytechnic}
\usage{
scale_color_vatech(palette = c("vatech"), alpha = 1, ...)
}
\description{
Scale Colors as Virginia Polytechnic
}
