% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_wm.R
\name{scale_color_wm}
\alias{scale_color_wm}
\title{Scale Colors as College of William & Mary}
\usage{
scale_color_wm(palette = c("wm"), alpha = 1, ...)
}
\description{
Scale Colors as College of William & Mary
}
