% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_american.R
\name{scale_fill_american}
\alias{scale_fill_american}
\title{Scale fill as American University}
\usage{
scale_fill_american(palette = c("american"), alpha = 1, ...)
}
\description{
Scale fill as American University
}
