% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_berkeley.R
\name{scale_fill_berkeley}
\alias{scale_fill_berkeley}
\title{Scale fill as UC Berkeley}
\usage{
scale_fill_berkeley(palette = c("berkeley"), alpha = 1, ...)
}
\description{
Scale fill as UC Berkeley
}
