% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_brownu.R
\name{scale_fill_brownu}
\alias{scale_fill_brownu}
\title{Scale Fill as Brown University}
\usage{
scale_fill_brownu(palette = c("brownu"), alpha = 1, ...)
}
\description{
Scale Fill as Brown University
}
