% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_georgetown.R
\name{scale_fill_georgetown}
\alias{scale_fill_georgetown}
\title{Scale Fill as Georgetown University}
\usage{
scale_fill_georgetown(palette = c("georgetown"), alpha = 1, ...)
}
\description{
Scale Fill as Georgetown University
}
