% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_harvard.R
\name{scale_fill_harvard}
\alias{scale_fill_harvard}
\title{Scale Fill as Harvard}
\usage{
scale_fill_harvard(palette = c("harvard"), alpha = 1, ...)
}
\description{
Scale Fill as Harvard
}
