% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_nyu.R
\name{scale_fill_nyu}
\alias{scale_fill_nyu}
\title{Scale Fill as New York University}
\usage{
scale_fill_nyu(palette = c("nyupal"), alpha = 1, ...)
}
\description{
Scale Fill as New York University
}
