% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_stanford.R
\name{scale_fill_stanford}
\alias{scale_fill_stanford}
\title{Scale Fill as Stanford University}
\usage{
scale_fill_stanford(palette = c("stanford"), alpha = 1, ...)
}
\description{
Scale Fill as Stanford University
}
