% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_uva.R
\name{scale_fill_uva}
\alias{scale_fill_uva}
\title{Scale Fill as University of Virginia}
\usage{
scale_fill_uva(palette = c("uvapal"), alpha = 1, ...)
}
\description{
Scale Fill as University of Virginia
}
