% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_vatech.R
\name{scale_fill_vatech}
\alias{scale_fill_vatech}
\title{Scale Fill as Virginia Polytechnic}
\usage{
scale_fill_vatech(palette = c("vatech"), alpha = 1, ...)
}
\description{
Scale Fill as Virginia Polytechnic
}
