% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_wm.R
\name{scale_fill_wm}
\alias{scale_fill_wm}
\title{Scale Fill as College of William & Mary}
\usage{
scale_fill_wm(palette = c("wm"), alpha = 1, ...)
}
\description{
Scale Fill as College of William & Mary
}
